feat: Add axios client configuration
This commit adds a new file, axios.ts, which contains the configuration for the axios HTTP client. The client is created with a base URL of 'http://localhost:3000', a timeout of 1000 milliseconds, and a 'Content-Type' header of 'application/json'. This configuration will be used for making HTTP requests throughout the application.