refactor: Add "Create Trip" button to header in RootLayoutNav
The RootLayoutNav component in _layout.tsx has been updated to include a "Create Trip" button in the header. This button allows users to easily navigate to the "create-trip" screen and create a new trip. This commit adds the necessary code for the button and improves the user experience by providing a convenient way to initiate trip creation.