feat: Add Actions component
The Actions component has been added to the repository. This component provides a set of actions that can be performed within a community. It enhances the user experience by allowing users to interact with the community and perform various actions such as adding a trip, editing the community, deleting the community, or leaving the community.