refactor: Update RootLayoutNav component
The RootLayoutNav component in _layout.tsx has been updated to include the Actions component in the community screen. This change enhances the user experience by providing a set of actions that can be performed within a community, such as adding a trip, editing the community, deleting the community, or leaving the community.