refactor: Add isAdmin and isMember properties to TCommunityDetail
This commit adds the `isAdmin` and `isMember` properties to the `TCommunityDetail` type in the `communityStore.ts` file. These properties will be used to determine if a user is an administrator or a member of a community. This enhancement improves the functionality of the community store by providing more information about the user's role within a community.