refactor: Add userVehicles property to TUserProfileWithUsername
The code changes in `authStore.ts` add the `userVehicles` property to the `TUserProfileWithUsername` type. This property represents an array of user vehicles, each containing an `id`, `year`, and `vehicle` object with properties such as `make`, `model`, and `vehicleType`. This addition enhances the user profile structure and allows for better management of user vehicle information in the application.