refactor: Update SelectInput component to support minimum search length
The code changes in `SelectInput.tsx` update the `SelectInput` component to support a minimum search length. The `minSearchLength` prop is added to specify the minimum number of characters required for a search to be performed. This enhancement improves the user experience by preventing unnecessary API calls for short search queries.