feat: Add createTrip functionality to Actions component
The code changes in `Actions.tsx` add the `createTrip` function to the Actions component. This function is triggered when the "Add Trip" button is pressed and navigates the user to the create trip page with the appropriate parameters. This feature enhances the application by allowing users to easily create new trips within the community.